style: make the workspace clippy-clean under CI's -D warnings
CI runs `cargo clippy --workspace --all-targets -- -D warnings`; newer clippy
(1.95) flags lints the code predates. Bring the whole workspace to zero
warnings/errors with no behavior changes:

- Auto-applied machine-fixable lints workspace-wide (unnecessary_cast,
  manual_range_contains, manual_is_multiple_of, redundant_pattern_matching,
  needless_return, question_mark, map_identity, useless_conversion,
  unnecessary_sort_by -> sort_by_key, ...).
- Removed ~40 redundant `let x = x;` rebindings in dispatch/{fs,net}.rs.
- Promoted constant `assert!`s to `const { assert!(...) }` (compile-time).
- Targeted, justified `#[allow]`s: genuinely-wide signatures (inject_signal),
  parallel-array index loops (page_table), a sharded-map field type, a
  once-per-run enum, and the duplicate-mod test-harness pattern.
- doc-comment list formatting (de-indented markers + paragraph separators).
- Added `OwnedHostMapping::is_empty`; dropped a dead `NOFILE_HARD` const (the
  live copy is a local const in dispatch/time.rs).

No-panic gate (unwrap/expect/panic denied) resolved without weakening it:
- carrick-image layer ordering returns the existing `OciBootstrapError`
  instead of `expect()` on a provably-non-empty bucket.
- fifo_beacon `STATE.lock().unwrap()` -> `.unwrap_or_else(|e| e.into_inner())`
  (recover from poison rather than panic).
- inotify: `let Some(record) = pop_front() else { break }` (front() already
  proved Some); fcntl F_SETOWN_EX uses explicit byte arrays (no try_into
  unwrap); pathcodec uses `unwrap_or(REPLACEMENT_CHARACTER)` for a
  provably-valid scalar.
- Intentionally-panicking test helpers (not `#[test]` fns, so
  allow-*-in-tests doesn't reach them) get a targeted `#[allow]`.

Verified: clippy --all-targets -D warnings, cargo fmt --check, cargo build
--workspace, cargo test --workspace --lib, and the consolidated integration
suite all green.

+++ crates/carrick-runtime/src/dispatch/fs/sendfile.rs
@@ -97,7 +97,7 @@ impl SyscallDispatcher {
             return None;
         }
         let st = unsafe { st.assume_init() };
-        Some((st.st_dev as i64, st.st_ino as u64))
+        Some((st.st_dev as i64, st.st_ino))
     }
 
     fn host_file_copy_info(&self, fd: i32) -> Option<HostFileCopyInfo> {
